STEP1: DatabaseParcelInformationPersistence 파일 분리
