+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Source/Service/DatabaseParcelInformationPersistence.swift
@@ -1,0 +1,19 @@
+//
+//  DatabaseParcelInformationPersistence.swift
+//  ParcelInvoiceMaker
+//
+//  Created by 홍승완 on 2024/03/04.
+//
+
+import Foundation
+
+protocol ParcelInformationPersistence {
+    func save(parcelInformation: ParcelInformation)
+}
+
+final class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
+    func save(parcelInformation: ParcelInformation) {
+        // 데이터베이스에 주문 정보 저장
+        print("발송 정보를 데이터베이스에 저장했습니다.")
+    }
+}
