+++ ParcelInvoiceMaker/ParcelInvoiceMaker/Source/Service/ParcelProcessor.swift
@@ -6,17 +6,6 @@
 
 import Foundation
 
-protocol ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation)
-}
-
-final class DatabaseParcelInformationPersistence: ParcelInformationPersistence {
-    func save(parcelInformation: ParcelInformation) {
-        // 데이터베이스에 주문 정보 저장
-        print("발송 정보를 데이터베이스에 저장했습니다.")
-    }
-}
-
 protocol ParcelOrderProcessable {
     func process(parcelInformation: ParcelInformation,
                  onComplete: (ParcelInformation) -> Void)
